Move `ViewData` and `ViewState` into a new `view_data` module
